Renamed Python interface C++ source files
As part of the reworking/restructuring of the Python interface renamed the majority of the C++ source files of the Python interface, because all of them reside in a dedicated "python" folder anyway, so having "Python" in their names is kind of redundant, and this way their somewhat long names got a bit shortened.
