--- python/server/PythonServerWrapper.h
@@ -1,6 +1,0 @@
-#ifndef __FreeOrion__PythonServerWrapper__
-#define __FreeOrion__PythonServerWrapper__
-
-void WrapServerAPI();
-
-#endif /* defined(__FreeOrion__PythonServerWrapper__) */
