+++ python/server/ServerWrapper.h
@@ -1,0 +1,6 @@
+#ifndef __FreeOrion__Python__ServerWrapper__
+#define __FreeOrion__Python__ServerWrapper__
+
+void WrapServerAPI();
+
+#endif /* defined(__FreeOrion__Python__ServerWrapper__) */
