+++ server/ServerApp.h
@@ -4,7 +4,7 @@
 
 #include "../util/Process.h"
 #include "../Empire/EmpireManager.h"
-#include "../python/server/PythonServerFramework.h"
+#include "../python/server/ServerFramework.h"
 #include "../network/ServerNetworking.h"
 #include "../universe/Universe.h"
 #include "../util/AppInterface.h"
